+++ src/server/game/AuctionHouse/AuctionHouseMgr.h
@@ -71,12 +71,12 @@ struct AuctionEntry
     uint32 auctioneer;                                      // creature low guid
     uint32 item_guidlow;
     uint32 item_template;
-    uint32 owner;
-    uint32 startbid;                                        //maybe useless
-    uint32 bid;
+    uint64 owner;
+    uint64 startbid;                                        //maybe useless
+    uint64 bid;
     uint32 buyout;
     time_t expire_time;
-    uint32 bidder;
+    uint64 bidder;
     uint32 deposit;                                         //deposit can be calculated only when creating auction
     AuctionHouseEntry const* auctionHouseEntry;             // in AuctionHouse.dbc
     uint32 factionTemplateId;
